Created a basic project
Created a basic C++ project with the Garry's Mod libraries, a premake5 file and everything else. (Yet to implement any of the actual Crypto++ libraries.)

Currently the only function is SHA256(string) which will return the same string you give as a parameter, or nil if you don't give it a parameter.