Delete old premake5 project & add new CMake project
This project was full of junk, so instead of trying to update it, I am just going to rewrite it from scratch.

The source for Libgcrypt and Crypto++ has been removed, along with all the premake5 and Visual Studio files.

CMake will be used going forward instead of premake5, as it is recommended on the Garry's Mod developer wiki and more commonly used.

The source code has been replaced with a basic example, however I do have working Crypto++ code that I will add at a later date.